+++ CHANGELOG.md
@@ -2,4 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- Property delegation to:
+  - `dagger.Lazy`
+  - `javax.inject.Provider`
 - Annotation processor generating Hilt modules binding primary implementations to their supertypes.